Extracted FitnessCalculator logics from Solution Added generics

+++ algen/algen-applications/src/main/java/it/red/algen/expressions/domain/ExprRawFitness.java
@@ -2,10 +2,15 @@ package it.red.algen.expressions.domain;
 
 import it.red.algen.domain.RawFitness;
 
+/**
+ * Absolute distance from target value
+ * @author red
+ *
+ */
 public class ExprRawFitness implements RawFitness {
-	public int distance;
+	public long distance;
 	
-	public ExprRawFitness(int distance){
+	public ExprRawFitness(long distance){
 		this.distance = distance;
 	}
